Move test to appropriate location; add handler test
